191022 | update 04_django_form, add 05_model_relation

+++ 04_django/04_django_form/articles/views.py
@@ -1,3 +1,4 @@
+import hashlib
 from IPython import embed
 from django.contrib.auth.decorators import login_required
 from django.shortcuts import render, redirect, get_object_or_404
@@ -8,7 +9,6 @@ from .forms import ArticleForm, CommentForm
 
 # Create your views here.
 def index(request):
-    # embed()
     # session 에 visits_num 키로 접근해 값을 가져온다.
     # visits_num은 기본적으로 존재하지 않은 키 이므로 키가 없다면(방문한적이 없다면) 0 값을 가져오도록 한다.
     visits_num = request.session.get('visits_num', 0)
@@ -33,7 +33,9 @@ def create(request):
         # 유효성 검증
         # form 이 유효한지 체크한다. (ex. blank=False와 같은 DB와 관련된 유효성 내용들)
         if form.is_valid():
-            article = form.save()
+            article = form.save(commit=False)
+            article.user = request.user # 원래는 article.user_id = request.user.pk
+            article.save()
             # form.cleaned_data 로 정제된 데이터를 받는다.
             # title = form.cleaned_data.get('title')
             # content = form.cleaned_data.get('content')
@@ -62,7 +64,7 @@ def detail(request, article_pk):
     # context = {'article': article,}
     # return render(request, 'articles/detail.html', context)
     article = get_object_or_404(Article, pk=article_pk)
-    comments = article.comment_set.all() # article의 모든 댓글
+    comments = article.comment_set.all() # article의 모든 댓글 (역참조시 : _set 사용)
     comment_form = CommentForm() # 댓글 form
     context = {'article': article, 'comment_form': comment_form, 'comments': comments,}
     return render(request, 'articles/detail.html', context)
@@ -72,28 +74,34 @@ def detail(request, article_pk):
 def delete(request, article_pk):
     if request.user.is_authenticated:
         article = get_object_or_404(Article, pk=article_pk)
-        article.delete()
+        if request.user == article.user:
+            article.delete()
+        else:
+            return redirect(article)
     return redirect('articles:index')
 
 
 @login_required
 def update(request, article_pk):
     article = get_object_or_404(Article, pk=article_pk)
-    if request.method == 'POST':
-        form = ArticleForm(request.POST, instance=article) # binding 작업 (instance : modelform에서 initial의 역할)
-        if form.is_valid(): # 유효성 검증
-            # article.title = form.cleaned_data.get('title') # 가져온 데이터(article.title)에 바꿔서 넣어준다.
-            # article.content = form.cleaned_data.get('content') # 가져온 데이터(article.content)에 바꿔서 넣어준다.
-            article = form.save()
-            return redirect(article)
+    if request.user == article.user:
+        if request.method == 'POST':
+            form = ArticleForm(request.POST, instance=article) # binding 작업 (instance : modelform에서 initial의 역할)
+            if form.is_valid(): # 유효성 검증
+                # article.title = form.cleaned_data.get('title') # 가져온 데이터(article.title)에 바꿔서 넣어준다.
+                # article.content = form.cleaned_data.get('content') # 가져온 데이터(article.content)에 바꿔서 넣어준다.
+                article = form.save()
+                return redirect(article)
+        else:
+            # embed()
+            # ArticleForm 을 초기화 (이전에 DB에 저장된 데이터를 넣어준 상태)
+            # form = ArticleForm(initial={'title': article.title, 'content': article.content}) # initial : 기존의 값을 가져온다(딕셔너리 형태로!)
+            # __dict__ : article 객체 데이터를 딕셔너리 자료형으로 변환
+            # form = ArticleForm(initial=article.__dict__)
+            form = ArticleForm(instance=article)
+            # 위와 같이 복잡한 한 줄은 매직 머서드(__dict__)로 줄여서 쓸 수 있다.
     else:
-        # embed()
-        # ArticleForm 을 초기화 (이전에 DB에 저장된 데이터를 넣어준 상태)
-        # form = ArticleForm(initial={'title': article.title, 'content': article.content}) # initial : 기존의 값을 가져온다(딕셔너리 형태로!)
-        # __dict__ : article 객체 데이터를 딕셔너리 자료형으로 변환
-        # form = ArticleForm(initial=article.__dict__)
-        form = ArticleForm(instance=article)
-        # 위와 같이 복잡한 한 줄은 매직 머서드(__dict__)로 줄여서 쓸 수 있다.
+        return redirect('articles:index')
     # 1. POST 방식일 때 넘어오는 form => 검증에 실패한 form(오류 메세지도 포함된 상태의 form)
     # 2. GET 방식일 때 넘어오는 form => 초기화된 form
     context = {'form': form, 'article': article,}
@@ -108,6 +116,7 @@ def comments_create(request, article_pk):
             # commit=False => 객체를 Create 하지만, db에 레코드는 작성하지 않는다.
             comment = comment_form.save(commit=False)
             comment.article_id = article_pk
+            comment.user = request.user
             comment.save()
     return redirect('articles:detail', article_pk)
 
@@ -116,7 +125,8 @@ def comments_create(request, article_pk):
 def comments_delete(request, article_pk, comment_pk):
     if request.user.is_authenticated:
         comment = get_object_or_404(Comment, pk=comment_pk)
-        comment.delete()
+        if request.user == comment.user:
+            comment.delete()
         return redirect('articles:detail', article_pk)
     return HttpResponse('You are Unauthorized', status=401)
     # 원래는 아래와 같이 작성
@@ -124,3 +134,17 @@ def comments_delete(request, article_pk, comment_pk):
     #     comment = get_object_or_404(Comment, pk=comment_pk)
     #     comment.delete()
     # return redirect('articles:detail', article_pk)
+
+def like(request, article_pk):
+    article = get_object_or_404(Article, pk=article_pk)
+    # 해당 게시글에 좋아요를 누른 사람들 중에서 현재 접속유저가 있다면 좋아요를 취소
+    if article.like_users.filter(pk=request.user.pk).exists():
+        # .get()은 없을 때 오류가 발생하므로 키가 없어도 오류(DoesNotExistError) 발생을 막기 위해 .filter()를 사용한다.
+        article.like_users.remove(request.user)
+    else:
+        article.like_users.add(request.user)
+    # if request.user in article.like_users.all():
+    #     article.like_users.remove(request.user) # 좋아요 취소
+    # else:
+    #     article.like_users.add(request.user) # 좋아요 선택
+    return redirect('articles:index')
